refactor: GSAM optimizer in gsam.py to sam.py
